ng,ngcore: refactor interactive REPL
